commit for test in server
config, controller/user, index.ts and server.ts have been changed
